+++ cmd/mistryd/worker.go
@@ -23,18 +23,22 @@ import (
 func (s *Server) Work(ctx context.Context, j *Job) (buildInfo *types.BuildInfo, err error) {
 	log := log.New(os.Stderr, fmt.Sprintf("[worker] [%s] ", j), log.LstdFlags)
 	start := time.Now()
+
+	// result cache
 	_, err = os.Stat(j.ReadyBuildPath)
 	if err == nil {
 		buildInfo, err := ReadJobBuildInfo(j.ReadyBuildPath, true)
 		if err != nil {
 			return nil, err
 		} else if buildInfo.ExitCode != 0 {
-			// previous build failed, remove its build dir to restart it
+			// Previous build failed, remove its build dir to
+			// restart it. We know it's not pointed to by a
+			// latest link since we only symlink successful builds
 			err = s.cfg.FileSystem.Remove(j.ReadyBuildPath)
 			if err != nil {
 				return buildInfo, workErr("could not remove existing failed build", err)
 			}
-		} else {
+		} else { // if a successful result exists already, return it
 			buildInfo.Cached = true
 			return buildInfo, err
 		}
@@ -44,38 +48,42 @@ func (s *Server) Work(ctx context.Context, j *Job) (buildInfo *types.BuildInfo, 
 	}
 
 	buildInfo = types.NewBuildInfo()
-	buildInfo.Path = filepath.Join(j.ReadyBuildPath, DataDir, ArtifactsDir)
-	buildInfo.TransportMethod = types.Rsync
-	buildInfo.Params = j.Params
-	buildInfo.StartedAt = j.StartedAt
-	buildInfo.URL = getJobURL(j)
+	j.BuildInfo = buildInfo
+	j.BuildInfo.Path = filepath.Join(j.ReadyBuildPath, DataDir, ArtifactsDir)
+	j.BuildInfo.TransportMethod = types.Rsync
+	j.BuildInfo.Params = j.Params
+	j.BuildInfo.StartedAt = j.StartedAt
+	j.BuildInfo.URL = getJobURL(j)
+	j.BuildInfo.Group = j.Group
 
 	added := s.jq.Add(j)
 	if added {
 		defer s.jq.Delete(j)
-	} else {
+	} else { // build coalescing
 		t := time.NewTicker(2 * time.Second)
-		log.Printf("Waiting for %s to complete...", j.PendingBuildPath)
+		defer t.Stop()
+		log.Printf("Coalescing with %s...", j.PendingBuildPath)
 		for {
 			select {
 			case <-ctx.Done():
-				err = workErr("context cancelled while waiting for pending build", nil)
+				err = workErr("context cancelled while coalescing", nil)
 				return
 			case <-t.C:
 				_, err = os.Stat(j.ReadyBuildPath)
 				if err == nil {
 					i, err := ExitCode(j)
 					if err != nil {
-						return buildInfo, err
+						return j.BuildInfo, err
 					}
-					buildInfo.ExitCode = i
-					buildInfo.Coalesced = true
-					return buildInfo, err
+					j.BuildInfo.ExitCode = i
+					j.BuildInfo.Coalesced = true
+					return j.BuildInfo, err
 				}
+
 				if os.IsNotExist(err) {
 					continue
 				} else {
-					err = workErr("could not wait for ready build", err)
+					err = workErr("could not coalesce", err)
 					return
 				}
 			}
@@ -98,13 +106,19 @@ func (s *Server) Work(ctx context.Context, j *Job) (buildInfo *types.BuildInfo, 
 		return
 	}
 
-	log.Printf("Creating new build directory...")
-	err = j.BootstrapBuildDir(s.cfg.FileSystem, log)
+	err = j.BootstrapBuildDir(s.cfg.FileSystem)
 	if err != nil {
 		err = workErr("could not bootstrap build dir", err)
+		return
 	}
 
-	// moves the pending directory to the ready one
+	err = persistBuildInfo(j)
+	if err != nil {
+		err = workErr("could not persist build info", err)
+		return
+	}
+
+	// move from pending to ready when finished
 	defer func() {
 		rerr := os.Rename(j.PendingBuildPath, j.ReadyBuildPath)
 		if rerr != nil {
@@ -116,9 +130,17 @@ func (s *Server) Work(ctx context.Context, j *Job) (buildInfo *types.BuildInfo, 
 			}
 		}
 
-		// When there are no errors the LatestBuildPath can be updated
-		// to point to the last build. Otherwise it remains unchanged.
+		// if the build was successful, symlink it to the 'latest'
+		// path
 		if err == nil {
+			// eliminate concurrent filesystem operations since
+			// they could result in a corrupted state (eg. if
+			// jobs of the same project simultaneously finish
+			// successfully)
+
+			s.pq.Lock(j.Project)
+			defer s.pq.Unlock(j.Project)
+
 			_, err = os.Lstat(j.LatestBuildPath)
 			if err == nil {
 				err = os.Remove(j.LatestBuildPath)
@@ -130,6 +152,7 @@ func (s *Server) Work(ctx context.Context, j *Job) (buildInfo *types.BuildInfo, 
 				err = workErr("could not stat the latest build link", err)
 				return
 			}
+
 			err = os.Symlink(j.ReadyBuildPath, j.LatestBuildPath)
 			if err != nil {
 				err = workErr("could not create latest build link", err)
@@ -137,9 +160,19 @@ func (s *Server) Work(ctx context.Context, j *Job) (buildInfo *types.BuildInfo, 
 		}
 	}()
 
-	if err != nil {
-		return
-	}
+	// populate j.BuildInfo.Err and persist it build_info file one last
+	// time
+	defer func() {
+		if err != nil {
+			j.BuildInfo.ErrBuild = err.Error()
+		}
+
+		err := persistBuildInfo(j)
+		if err != nil {
+			err = workErr("could not persist build info", err)
+			return
+		}
+	}()
 
 	for k, v := range j.Params {
 		err = ioutil.WriteFile(filepath.Join(j.PendingBuildPath, DataDir, ParamsDir, k), []byte(v), 0644)
@@ -166,22 +199,22 @@ func (s *Server) Work(ctx context.Context, j *Job) (buildInfo *types.BuildInfo, 
 		}
 	}()
 
-	biJSON, err := json.Marshal(buildInfo)
-	if err != nil {
-		err = workErr("could not serialize build info", err)
-		return
-	}
-	err = ioutil.WriteFile(j.BuildInfoFilePath, biJSON, 0666)
-	if err != nil {
-		err = workErr("could not write build info to file", err)
-		return
-	}
-
 	client, err := docker.NewEnvClient()
 	if err != nil {
 		err = workErr("could not create docker client", err)
 		return
 	}
+	defer func() {
+		derr := client.Close()
+		errstr := "could not close docker client"
+		if derr != nil {
+			if err == nil {
+				err = fmt.Errorf("%s; %s", errstr, derr)
+			} else {
+				err = fmt.Errorf("%s; %s | %s", errstr, derr, err)
+			}
+		}
+	}()
 
 	err = j.BuildImage(ctx, s.cfg.UID, client, out, j.Rebuild, j.Rebuild)
 	if err != nil {
@@ -190,41 +223,29 @@ func (s *Server) Work(ctx context.Context, j *Job) (buildInfo *types.BuildInfo, 
 	}
 
 	var outErr strings.Builder
-	buildInfo.ExitCode, err = j.StartContainer(ctx, s.cfg, client, out, &outErr)
-
+	j.BuildInfo.ExitCode, err = j.StartContainer(ctx, s.cfg, client, out, &outErr)
 	if err != nil {
 		err = workErr("could not start docker container", err)
 		return
 	}
 
-	biJSON, err = json.Marshal(buildInfo)
-	if err != nil {
-		err = workErr("could not serialize build info", err)
-		return
-	}
-	err = ioutil.WriteFile(j.BuildInfoFilePath, biJSON, 0666)
-	if err != nil {
-		err = workErr("could not write build info to file", err)
-		return
-	}
-
-	// fill the buildInfo Log from the freshly written log
 	err = out.Sync()
 	if err != nil {
 		err = workErr("could not flush the output log", err)
 		return
 	}
 
-	finalLog, err := ReadJobLogs(j.PendingBuildPath)
+	stdouterr, err := ReadJobLogs(j.PendingBuildPath)
 	if err != nil {
 		err = workErr("could not read the job logs", err)
 		return
 	}
 
-	buildInfo.Log = string(finalLog)
-	buildInfo.ErrLog = outErr.String()
+	j.BuildInfo.ContainerStdouterr = string(stdouterr)
+	j.BuildInfo.ContainerStderr = outErr.String()
+	j.BuildInfo.Duration = time.Now().Sub(start).Truncate(time.Millisecond)
 
-	log.Println("Finished after", time.Now().Sub(start).Truncate(time.Millisecond))
+	log.Println("Finished after", j.BuildInfo.Duration)
 	return
 }
 
@@ -275,4 +296,20 @@ func workErr(s string, e error) error {
 		s += "; " + e.Error()
 	}
 	return errors.New(s)
+}
+
+// persistBuildInfo persists the JSON-serialized version of j.BuildInfo
+// to disk.
+func persistBuildInfo(j *Job) error {
+	// we don't want to persist the whole build logs in the build_info file
+	bi := *j.BuildInfo
+	bi.ContainerStdouterr = ""
+	bi.ContainerStderr = ""
+
+	out, err := json.Marshal(bi)
+	if err != nil {
+		return err
+	}
+
+	return ioutil.WriteFile(j.BuildInfoFilePath, out, 0666)
 }